set cursorline
I'm doing this for a terrible reason, because in some source files (go
and javascript to name a couple), the trailing whitespace is highlighted
as I type. This is fucking annoying and I can't figure out how to
disable it and highlighting the current cursor line gets rid of it on
the line I'm currently working on.